fix swap error

+++ scripts/cleanup-14.04.sh
@@ -26,13 +26,6 @@ rm -rfv /usr/src/vboxguest* /usr/src/virtualbox-ose-guest*
 # Clean any Python bytecode
 find / -type f -name "*.pyc" -delete
 
-# Whiteout root
-echo "Whiteouting /"
-count=`df --sync -kP / | tail -n1  | awk -F ' ' '{print $4}'`;
-count=$((count -= 1))
-dd if=/dev/zero of=/tmp/whitespace bs=1024 count=$count;
-rm /tmp/whitespace;
- 
 # Whiteout /boot
 echo "Whiteouting /boot"
 count=`df --sync -kP /boot | tail -n1 | awk -F ' ' '{print $4}'`;
@@ -41,13 +34,17 @@ dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count;
 rm /boot/whitespace;
  
 # Whiteout swap 
-echo "Whiteouting swap"
-swappart=`cat /proc/swaps | tail -n1 | awk -F ' ' '{print $1}'`
-swapoff $swappart;
-dd if=/dev/zero of=$swappart;
-mkswap $swappart;
-swapon $swappart;
-
+echo "Whiteouting swaps"
+if [ $(cat /proc/swaps | wc -l) -gt 1 ]; then
+  for swappart in `cat /proc/swaps | tail -n +2 | awk -F ' ' '{print $1}'`; do
+    oldUUID="$(blkid $swappart | awk '{print $2}' | cut -d '"' -f 2)"
+    swapoff $swappart;
+    dd if=/dev/zero of=$swappart;
+    mkswap -U $oldUUID $swappart;
+    swapon $swappart;
+  done
+  unset oldUUID
+fi
 # Zero out the rest of the free space using dd.
 echo "Zero out the free space to aid VM compression"
 dd if=/dev/zero of=/EMPTY bs=1M
